backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
faster-whisper>=1.0.3
av>=12.3.0
openai>=1.51.0
pymongo>=4.10.1
PyJWT[crypto]>=2.9.0
chromadb>=0.5.0
httpx>=0.27.0
langchain>=0.2.0
langchain-community>=0.2.0
tiktoken>=0.7.0
boto3>=1.34.0
